Resolve S13P21A407-96 "Feature/android/홈 서버 연동 "

See merge request s13-fintech-finance-sub1/S13P21A407!184

+++ pocketc_android/app/src/main/java/com/example/irumi/domain/entity/main/StreakEntity.kt
@@ -3,6 +3,6 @@ package com.example.irumi.domain.entity.main
 data class StreakEntity(
     val date: String,              // yyyy-MM-dd
     val missionsCompleted: Int,
-    val spending: Int,
+    val spending: Long,
     val isActive: Boolean
 )
